+++ PCE_Web/PCE_Web/Classes/ExceptionsManager.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using PCE_Web.Models;
 using PCE_Web.Tables;
 
@@ -13,10 +14,39 @@ namespace PCE_Web.Classes
             _pceDatabaseContext = pceDatabaseContext;
         }
 
-        public void WriteLoggedExceptions(string message, string source, string stackTrace, string date);
+        public void WriteLoggedExceptions(string message, string source, string stackTrace, string date)
+        {
+            var result = _pceDatabaseContext.SavedExceptions.SingleOrDefault(column => column.Date == date && column.Message == message && column.Source == source && column.StackTrace == stackTrace);
 
-        public void DeleteLoggedExceptions(int id);
+            if (result == null)
+            {
+                var savedExceptions = new SavedExceptions()
+                {
+                    Message = message,
+                    Source = source,
+                    StackTrace = stackTrace,
+                    Date = date
+                };
+                _pceDatabaseContext.SavedExceptions.Add(savedExceptions);
+                _pceDatabaseContext.SaveChanges();
+            }
+        }
 
-        public List<Exceptions> ReadLoggedExceptions();
+        public void DeleteLoggedExceptions(int id)
+        {
+            var result = _pceDatabaseContext.SavedExceptions.SingleOrDefault(column => column.SavedExceptionId == id);
+            if (result != null)
+            {
+                _pceDatabaseContext.SavedExceptions.Remove(result);
+                _pceDatabaseContext.SaveChanges();
+            }
+        }
+
+        public List<Exceptions> ReadLoggedExceptions()
+        {
+            var exceptions = _pceDatabaseContext.SavedExceptions.Where(row => row.SavedExceptionId > 0)
+                .Select(column => new Exceptions { Id = column.SavedExceptionId, Date = column.Date, Message = column.Message, StackTrace = column.StackTrace, Source = column.Source }).ToList();
+            return exceptions;
+        }
     }
 }
