+++ PCE_Web/PCE_Web/Classes/ReportsManager.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using PCE_Web.Models;
 using PCE_Web.Tables;
 
@@ -13,14 +15,55 @@ namespace PCE_Web.Classes
             _pceDatabaseContext = pceDatabaseContext;
         }
 
-        public void WriteReports(string email, string report);
+        public void WriteReports(string email, string report)
+        {
+            if (report != null)
+            {
+                var newReport = new Reports()
+                {
+                    Email = email,
+                    Comment = report,
+                    Date = DateTime.UtcNow.ToString(),
+                    Solved = 0
+                };
+                _pceDatabaseContext.Reports.Add(newReport);
+                _pceDatabaseContext.SaveChanges();
+            }
+        }
 
-        public List<Report> ReadReports(string email, int solvedId);
+        public List<Report> ReadReports(string email, int solvedId)
+        {
+            var comments = _pceDatabaseContext.Reports.Where(column => column.Email == email && column.Solved == solvedId).Select(column => new Report { Comment = column.Comment, ID = column.ReportsId, Date = column.Date, Email = column.Email }).ToList();
+            return comments;
+        }
 
-        public void DeleteReports(int id);
+        public void DeleteReports(int id)
+        {
+            var result = _pceDatabaseContext.Reports.SingleOrDefault(column => column.ReportsId == id);
+            if (result != null)
+            {
+                _pceDatabaseContext.Reports.Remove(result);
+                _pceDatabaseContext.SaveChanges();
+            }
+        }
 
-        public bool IsReported(string email);
+        public bool IsReported(string email)
+        {
+            var item = _pceDatabaseContext.Reports.Where(column => column.Email == email).Select(column => new Reports { Email = column.Email, Comment = column.Comment, Solved = column.Solved, Date = column.Date }).ToList();
+            if (item.Count > 0)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
 
-        public void MarkAsSolved(int id);
+        public void MarkAsSolved(int id)
+        {
+            _pceDatabaseContext.Reports.Where(column => column.ReportsId == id && column.Solved == 0).ToList().ForEach(column => column.Solved = 1);
+            _pceDatabaseContext.SaveChanges();
+        }
     }
 }
